Added solution for day 1 in ocaml
The solution in Ocaml so far has some quite rudimentary (slow, unoptimized)
ways for calculating the answer.
There is some optimization from part1a to part1b, but we could still
make it so we exit the for-loop earlier in part1b (or use something other
than a for-loop, like while-loop or recursive function.).